+++ x-pack/plugin/sql/sql-action/build.gradle
@@ -3,8 +3,7 @@
  */
 apply plugin: 'elasticsearch.build'
 
-description = 'Request and response objects shared by the cli, jdbc ' +
-  'and the Elasticsearch plugin'
+description = 'Request and response objects shared by the Elasticsearch plugin and qa tests'
 
 dependencies {
   /* We'd like to just depend on xcontent but there are some bits of
